pyproject.toml
[project]
name = "crypto-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.3",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "flower>=2.0.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
